Added 1.12.2 pack version
This version is fully compatible with 1.12 and has almost all features of the main pack. The only missing feature is the trapped chest model, which is due to trapped chests not having a discrete model like they do in later versions. Instead the chest just has a bloodied appearance, separating it from normal chests.